Comment: change "replyTo" type from int to long

+++ src/main/kotlin/kmtt/models/comment/Comment.kt
@@ -74,7 +74,7 @@ data class Comment(
     val entry: Entry? = null,
 
     @JsonNames("replyTo")
-    val replyTo: Int? = null,
+    val replyTo: Long? = null,
 
     @JsonNames("isFavorited")
     val isFavorited: Boolean? = null,
